svc-rabbitmq/pyproject.toml
[tool.poetry]
name = "svc_rabbitmq"
version = "1.0.0"
description = ""
authors = ["blueking <blueking@tencent.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.9"
django = "==3.2.12"
pymysql = "==1.0.2"
gunicorn = "==20.1.0"
paas-service = "==1.1.4"
djangorestframework = "==3.12.4"
requests = "==2.21.0"
gevent = "==21.12.0"
cryptography = ">=3.0.0,<4.0.0"
sentry-sdk = "==0.19.4"
jsonfield = "==3.1.0"
whitenoise = "==5.3.0"
django-q = "1.3.9"
django-redis = "5.2.0"
arrow = "1.2.2"
asn1crypto = "1.5.1"
blue-krill = ">=1.0.15"
django-picklefield = "3.0.1"
python-dateutil = "2.8.2"
prometheus_client = "0.14.1"
AMQPStorm = "2.10.4"
pydantic = "==1.8"
django-prometheus = "2.2.0"
django-environ = "^0.4.5"

[tool.poetry.dev-dependencies]
flake8 = "*"
pytest-django = "*"
pytest = "*"
requests-mock = "1.5.2"
ipython = "*"
ipdb = {version = "^0.13.2", allow-prereleases = true}

[[tool.poetry.source]]
name = "pypi-tencent-mirror"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
default = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
